Fix ESP-IDF 5.x build compatibility for GS firmware
- Fix unified_rx API to match actual WfbReceiver/FrameAssembler/FecDecoder interfaces
- Add stddef.h include for size_t in unified_rx.h
- Fix callback signatures to match WfbDataCallback/FrameDecodedCallback
- Fix wfb_bridge API calls (set_data_callback, get_stats, init)
- Update wfb_crypto to use mbedtls/ecdh.h instead of non-existent curve25519.h
- Fix format specifiers for cross-platform types (uint32_t, size_t)
- Update USB NCM implementation with TinyUSB NCM callbacks
- Add idf_component.yml files for esp_tinyusb dependency
- Enable ChaCha20-Poly1305 in sdkconfig.defaults for WFB-NG encryption
- Fix web file embedding in gs_main CMakeLists.txt
- Fix video_ws_server ws_open_callback return type
- Add missing FreeRTOS/heap_caps includes to main.cpp

Compilation now succeeds; linker issues remain for FEC decoder and mbedtls symbols.

+++ gs_firmware_esp32s3/main/main.cpp
@@ -12,8 +12,12 @@
  */
 
 #include <stdio.h>
+#include <inttypes.h>
 #include "esp_log.h"
 #include "esp_system.h"
+#include "esp_heap_caps.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
 #include "gs_main.h"
 
 static const char* TAG = "main";
@@ -25,11 +29,11 @@ extern "C" void app_main(void)
     ESP_LOGI(TAG, "=================================");
 
     // Print memory info
-    ESP_LOGI(TAG, "Free heap: %" PRIu32 " bytes", esp_get_free_heap_size());
-    ESP_LOGI(TAG, "Free internal: %" PRIu32 " bytes",
-             heap_caps_get_free_size(MALLOC_CAP_INTERNAL));
-    ESP_LOGI(TAG, "Free PSRAM: %" PRIu32 " bytes",
-             heap_caps_get_free_size(MALLOC_CAP_SPIRAM));
+    ESP_LOGI(TAG, "Free heap: %lu bytes", (unsigned long)esp_get_free_heap_size());
+    ESP_LOGI(TAG, "Free internal: %lu bytes",
+             (unsigned long)heap_caps_get_free_size(MALLOC_CAP_INTERNAL));
+    ESP_LOGI(TAG, "Free PSRAM: %lu bytes",
+             (unsigned long)heap_caps_get_free_size(MALLOC_CAP_SPIRAM));
 
     // Initialize ground station with default config
     gs_config_t config = {
